add converter.ts
