+++ source/binary.ts
@@ -47,23 +47,6 @@ const readonlyArrayNumber: typeExpr.TypeExpr = typeExpr.withTypeParameter(
   [typeExpr.typeNumber]
 );
 
-/**
- * UnsignedLeb128で表現されたバイナリに変換する
- */
-const encodeUInt32 = (num: number): ReadonlyArray<number> => {
-  num = Math.floor(Math.max(0, Math.min(num, 2 ** 32 - 1)));
-  const numberArray = [];
-  while (true) {
-    const b = num & 0x7f;
-    num = num >>> 7;
-    if (num === 0) {
-      numberArray.push(b);
-      return numberArray;
-    }
-    numberArray.push(b | 0x80);
-  }
-};
-
 const encodeUInt32Name = ["decodeUInt32"];
 export const encodeUInt32Var = expr.localVariable(encodeUInt32Name);
 const mathObject = expr.globalVariable("Math");
@@ -130,26 +113,6 @@ export const encodeUInt32Code = expr.functionWithReturnValueVariableDefinition(
     ])
   ]
 );
-
-/**
- * UnsignedLeb128で表現されたバイナリをnumberの32bit符号なし整数の範囲の数値にに変換する
- */
-export const decodeUInt32 = (
-  index: number,
-  binary: Uint8Array
-): { result: number; nextIndex: number } => {
-  let result = 0;
-
-  for (let i = 0; i < 5; i++) {
-    const b = binary[index + i];
-
-    result |= (b & 0x7f) << (7 * i);
-    if ((b & 0x80) === 0 && 0 <= result && result < 2 ** 32 - 1) {
-      return { result, nextIndex: index + i + 1 };
-    }
-  }
-  throw new Error("larger than 32-bits");
-};
 
 const decodeUInt32Name = ["decodeUInt32"];
 export const decodeUInt32Var = expr.localVariable(decodeUInt32Name);
@@ -340,30 +303,12 @@ export const requestObjectTypeToTypeAlias = (
   };
 };
 
-export const decodeString = (
-  index: number,
-  binary: Uint8Array
-): { result: string; nextIndex: number } => {
-  const length = decodeUInt32(index, binary);
-  return {
-    result: new TextDecoder().decode(
-      binary.slice(
-        index + length.nextIndex,
-        index + length.nextIndex + length.result
-      )
-    ),
-    nextIndex: index + length.nextIndex + length.result
-  };
-};
-
 const decodeStringName = ["decodeString"];
 export const decodeStringVar = expr.localVariable(decodeStringName);
 
 /**
  * バイナリからstringに変換するコード
- * ブラウザではグローバルの
- * @param textDecoderExpr
- * @param integerDecoderIndex
+ * ブラウザではグローバルのTextDecoderを使い、node.jsではutilのTextDecoderを使う
  */
 export const decodeStringCode = (isBrowser: boolean): expr.Statement =>
   expr.functionWithReturnValueVariableDefinition(
@@ -582,25 +527,3 @@ export const typeToTypeExprVar = <
     }
   }
 };
-
-/*
-  {
-  const patternIdAndNextIndex = numberFromUnsignedLeb128(index, binary);
-  const patternId = patternIdAndNextIndex.result;
-  if (patternId === 0) {
-    const nameAndNextIndex = stringDecoder(
-      patternIdAndNextIndex.nextIndex,
-      binary
-    );
-    const ageAndNextIndex = numberFromUnsignedLeb128(
-      nameAndNextIndex.nextIndex,
-      binary
-    );
-    const result = createUser(
-      nameAndNextIndex.result,
-      ageAndNextIndex.nextIndex
-    );
-  }
-}
-
-*/
